Split engine/claude_test.go into focused test files
Move TestBuildPrompt_*/TestBuildCommentReviewPrompt_* to build_prompt_test.go,
TestInvokeClaude_*/TestRealClaudeInvoker_*/TestDirectInvocation/TestSaveDebugLog*
to invoke_claude_test.go, TestParseClaudeJSON_*/TestCheckCompletion_*/TestTokenUsage*
to parse_claude_test.go, TestSaveSessionID*/TestSessionDir/TestLogDir/TestFormatStatsFooter
to session_test.go. Delete claude_test.go.
